REF: DataFrame.mask tests (#33219)
